feat: move jinja rendering in jinja module
